+++ extension/src/astObjects/sexpression.ts
@@ -1,426 +1,11 @@
-import { closeParenStyle, maximumLineChars, longListFormatStyle, indentSpaces } from './fmtconfig';
+import { closeParenStyle, indentSpaces, longListFormatStyle, maximumLineChars } from '../format/fmtconfig';
 import { isInternalAutoLispOp } from '../completion/autocompletionProvider';
-import { Position, Range } from 'vscode';
-
-// General purpose test for basic known primitive values; Including: T, NIL, Number, (single or multi-line) Strings & Comments
-export const primitiveRegex = /^([\(\)\'\.]|"[\s\S]*"|;[\s\S]*|'?[tT]|'?[nN][iI][lL]|'?-?\d+[eE][+-]?\d+|-?\d+|-?\d+\.\d+)$/;
-const primitiveGlyphs = ['\'', '(', ')', '.', ';']; //, '']; //, null, undefined];
-
-// This interface is intended to let us work in more generic ways without direct context of LispAtom|LispContainer
-export interface ILispFragment {
-    symbol: string;
-    line: number;
-    column: number;
-    flatIndex: number;
-    commentLinks?: Array<number>;
-    hasGlobalFlag?: boolean;
-
-    
-    readonly body?: LispContainer|undefined;
-    
-    isLispFragment(): boolean;
-    equal(atom: ILispFragment): boolean;
-    symbLine(last?: boolean): number;
-    length(): number;
-    isLineComment(): boolean;
-    isComment(): boolean;
-    isRightParen(): boolean;
-    isLeftParen(): boolean;
-    isPrimitive(): boolean;
-    contains(pos: Position): boolean;
-    getRange(): Range;
-}
-
-// Represents the most fundamental building blocks of a lisp document
-export class LispAtom implements ILispFragment {
-    public symbol: string;
-    protected _line: number;
-    protected _column: number;
-
-    get line(): number { return this._line; }
-    set line(value) { this._line = value; }
-    get column(): number { return this._column; }
-    set column(value) { this._column = value; }
-
-    // These 3 fields exist to support comment driven intelligence. Creation of Symbol mappings is an expensive operation
-    // and these 2 fields prevent ~80% of the required overhead when working in concert with the highly efficient parser
-    public flatIndex: number;
-    public commentLinks?: Array<number>;
-    public hasGlobalFlag?: boolean;
-
-    constructor(line: number, column: number, sym: string, flatIdx = -1) {
-        this._line = line;
-        this._column = column;
-        this.symbol = sym;
-        this.flatIndex = flatIdx;
-    }
-
-
-    // Does a simple comparison between 2 ILispFragments without the reference problem
-    equal(atom: ILispFragment): boolean {
-        return JSON.stringify(this) === JSON.stringify(atom);
-    }
-
-
-    // Determines if this LispAtom or its derived type can be used as an ILispFragment
-    isLispFragment(): boolean {
-        if (this instanceof Sexpression) {
-            return false;
-        } else {
-            return true;
-        }
-    }
-
-
-    // returns the start or ending line of the LispAtom depending on the boolean flag
-    symbLine(last: boolean = true): number {
-        if (last) {
-            let internalLines = 0;
-            if (this.symbol.startsWith(';|')) {
-                for (let i = 0; i < this.symbol.length; i++) {
-                    if (this.symbol.charAt(i) === '\n') { // it can handle the \r\n and \n 
-                        internalLines++;
-                    }
-                }
-            }
-            return this.line + internalLines;
-        }
-        else {
-            return this.line; 
-        }
-    }
-
-
-    // Returns the length of the LispAtom's text value
-    length(): number {
-        return this.symbol.length;
-    }
-
-
-    // Tests if the LispAtom is representing a single-line comment
-    isLineComment(): boolean {
-		return this.symbol.startsWith(';') && !this.symbol.startsWith(';|');
-    }
-	
-
-    // Tests if the LispAtom is representing any type of comment
-    isComment(): boolean {
-		return this.symbol.startsWith(';');
-    }
-	
-
-    // Tests if the LispAtom is representing a structural closing parenthesis
-    isRightParen(): boolean {
-		return this.symbol === ')';
-    }
-	
-
-    // Tests if the LispAtom is representing a structural opening parenthesis
-    isLeftParen(): boolean {
-		return this.symbol === '(';
-    }
-
-    isPrimitive(): boolean {
-        // if (!this['atoms']) {
-        //     return primitiveRegex.test(this.symbol);
-        // }
-        // return false;
-        return primitiveGlyphs.indexOf(this.symbol[0]) > -1
-               || primitiveRegex.test(this.symbol);
-    }
-
-    // Returns true if this LispAtom encapsulates the provided Position
-    contains(position: Position): boolean {
-        return this.getRange().contains(position);
-    }
-
-
-    // Gets the full range of the LispAtom and is capable of handling multi line strings or comments
-    getRange(): Range {
-        let cLine = this.line;
-        let cColm = this.column;
-        const begin: Position = new Position(cLine, cColm);
-        for (let i = 0; i < this.symbol.length; i++) {
-            const ch = this.symbol[i];
-            if (ch === '\n') {
-                cLine += 1;
-                cColm = 0;
-            } else {
-                cColm += 1;
-            }
-        }
-        const close: Position = new Position(cLine, cColm);        
-        return new Range(begin.line, begin.character, close.line, close.character);
-    }
-
-}
-
-
-// A utility class to hold and work with a collection of LispAtom|LispContainers
-export class LispContainer extends LispAtom {
-    atoms: Array<ILispFragment>;
-    linefeed: string;
-    userSymbols?: Map<string, Array<number>>; // this would only show up on document root LispContainers
-
-    // pass through getter for the ILispFragment interface
-    get body(): LispContainer { return this; }
-    get line(): number { return this.userSymbols ? 0 : this.getFirstAtom().line; }
-    set line(value) { }   // setter exists only to satisfy contract
-    get column(): number { return this.userSymbols ? 0 : this.getFirstAtom().column; }
-    set column(value) { } // setter exists only to satisfy contract
-
-    // LispContainer constructor defaults to a clearly uninitialized state
-    constructor(lineFeed: string = '\n') {
-        super(-1,-1,'');
-        this.atoms = [];
-        this.linefeed = lineFeed;
-    }
-
-
-    // Returns a crude version of the text a LispAtom|LispContainer represents
-    static getRawText(lispObj: ILispFragment): string {
-        if (lispObj instanceof LispContainer) {            
-            let ret = '';
-            for (let atom of lispObj.atoms) {
-                ret += LispContainer.getRawText(atom) + ' ';
-            }
-            return ret;
-        } else {
-            return lispObj.symbol;
-        }
-    }
-
-
-    // Adds one or more ILispFragment's to the LispContainer Atoms
-    addAtom(...items: ILispFragment[]) {
-        items.forEach(f => {
-            if (f?.isLispFragment()) {
-                this.atoms.push(f);
-            }
-        });
-    }
-
-
-    // Finds a child LispAtom if it legitimately touches/contains the Position
-    getAtomFromPos(loc: Position): LispAtom {
-        if (!this.contains(loc)) {
-            return null;
-        } else {
-            for (var i = 0; i < this.atoms.length; i++) {
-                const lispObj = this.atoms[i];
-                if (lispObj.contains(loc)) {
-                    if (lispObj instanceof LispContainer) {
-                        return lispObj.getAtomFromPos(loc);                        
-                    } else if (lispObj instanceof LispAtom) {
-                        return lispObj;
-                    }
-                }
-            }
-        }
-        return null;
-    }
-
-
-    // Returns the total length of characters within a LispContainer minus any LispAtom separated whitespace
-    length(): number {
-        let res = 0;
-        this.atoms.forEach(atom => {
-            res += atom.length();
-        });
-        return res;
-    }
-
-
-    // returns the start or ending line of the LispContainer depending on the boolean flag
-    symbLine(last: boolean = true): number {
-        if (this.atoms.length === 0) {
-            return -1;
-        }
-
-        if (last) {
-            const lastAtom = this.atoms[this.atoms.length - 1];
-            return lastAtom.symbLine();
-        } else {
-            return this.atoms[0].symbLine();
-        }
-    }
-
-
-    // gets the total length of root atoms this LispContainer holds
-    size(): number {
-        return this.atoms.length;
-    }
-
-
-    // Finds a child LispContainer if it legitimately touches/contains the Position
-    getExpressionFromPos(loc: Position, parent: boolean = false): LispContainer {
-        let result: LispContainer = null;
-        if (this.contains(loc)) {
-            this.atoms.forEach(lispObj => {
-                if (lispObj.contains(loc) && lispObj instanceof LispContainer) {
-                    result = lispObj.getExpressionFromPos(loc, parent) ?? (parent ? this : lispObj);
-                }
-            });
-        }
-        return result;
-    }
-
-
-    // Finds the parent LispContainer of any existing ILispFragment, typically used with a primary Container like Defun
-    getParentOfExpression(child: ILispFragment): LispContainer {        
-        const pos = new Position(child.line, child.column);
-        return this.getExpressionFromPos(pos, true);
-    }
-
-
-    // Gets a range representing the full LispContainer, especially useful for TextDocument.getText()
-    getRange(): Range {
-        const begin = this.getFirstAtom().getRange().start;
-        const close = this.getLastAtom().getRange().end;
-        return new Range(begin, close);
-    }
-
-    private getFirstAtom(): ILispFragment {
-        const tail = this.atoms[0];
-        return tail.body?.getFirstAtom() ?? tail;
-    }
-    private getLastAtom(): ILispFragment {
-        const tail = this.atoms[this.atoms.length - 1];
-        return tail.body?.getLastAtom() ?? tail;
-    }
-
-
-
-    // This version was necessary to properly alternate over SETQ Name vs Value 
-    private isValidForSetq(atom: ILispFragment): boolean {
-        return !atom.isComment() && !['\'', '(', ')', '.'].includes(atom.symbol) && (atom instanceof LispContainer || atom.symbol.trim().length > 0);
-    }    
-    // This is general purpose utility to make sure primitives such as strings, numbers and decorations are not evaluated
-    private notNumberStringOrProtected(atom: ILispFragment): boolean {        
-        return !atom.isComment() && !['\'', '(', ')', '.'].includes(atom.symbol) && !(/^".*"$/.test(atom.symbol)) && !(/^\'{0,1}\-{0,1}\d+[eE\-]{0,2}\d*$/.test(atom.symbol));
-    }
-
-
-    // This is an iteration helper that makes sure only useful LispAtom/LispContainers's get used for data collection purposes.
-    nextKeyIndex(currentIndex: number, forSetq?: boolean): number {
-        let index = currentIndex + 1;
-        if (index < this.atoms.length) {
-            const atom = this.atoms[index];
-            const flag = forSetq ? this.isValidForSetq(atom) : this.notNumberStringOrProtected(atom);
-            if (atom instanceof LispAtom && flag) {
-                return index;               
-            } else {
-                return this.nextKeyIndex(index, forSetq);
-            }
-        } else {
-            return -1;
-        }
-    }
-
-
-    // Returns a meaningful value from the LispContainer, this is useful for avoiding comments an structural symbols
-    getNthKeyAtom(significantNth: number): ILispFragment {
-        let num = 0;        
-        for (let i = 0; i <= significantNth; i++) {
-            num = this.nextKeyIndex(num, true);
-        }          
-        if (num < this.atoms.length && num >= 0) {
-            return this.atoms[num];
-        } else {
-            return null;
-        }
-    }
-
-
-    // Used a lot like a DOM selector to navigate/extract values from LispContainer
-    findChildren(regx: RegExp, all: boolean): LispContainer[] {
-        let result: Array<LispContainer> = [];       
-        let index = this.nextKeyIndex(0);
-        const lispObj = this.atoms[index];
-        if (!lispObj) {
-            return result;
-        } else if (regx.test(lispObj.symbol) === true) {
-            result.push(this);
-            if (all === true) {
-                this.atoms.filter(f => f.body).forEach((atom: ILispFragment) => {
-                    result = result.concat(atom.body.findChildren(regx, all));
-                });
-            }
-        } else {
-            this.atoms.filter(f => f.body).forEach((atom: ILispFragment) => {
-                result = result.concat(atom.body.findChildren(regx, all));
-            });
-        }
-        return result;
-    }
-
-    // Performance Note: the LispContainer parser takes ~700ms to create the tree from a 12.5mb (20K Lines/1.4M LispAtoms)
-    //                   file. In contrast, the older parsing used on formatting before it received some fixes was taking
-    //                   more than a minute and the revised version is still taking ~7000ms.
-    //                   To flatten the same 12.5mb tree for linear traversal takes ~40ms. The flattening process is
-    //                   considered to be an instantaneous operation even under extreme and highly improbable situations.
-    flatten(into?: Array<LispAtom>): Array<LispAtom> {
-        if (!into) {
-            into= [];
-        }
-        this.atoms.forEach(item => {
-            if (item instanceof LispContainer) {                
-                item.flatten(into);
-            } else if (item instanceof LispAtom) {
-                into.push(item);
-            }
-        });
-        return into;
-    }
-
-    
-    // This is only used for verification of the LispContainer parser, but could have future formatting usefullness.
-    // Note1: test files cannot contain lines with only random whitespace or readable character lines that include
-    //        trailing whitespace because a LispContainer has no context to reproduce that. 
-    // Note2: tabs between atoms will be replaced with spaces and thus are not supported in test files.
-    asText(context?: IContainerStringCompilerContext): string {
-        let isRoot = false;
-        if (context === undefined) {
-            context = { result: '', line: 0, column: 0 };
-            isRoot = true;
-        }
-        this.atoms.forEach(item => {
-            while (context.line < item.line) {
-                context.result += this.linefeed;
-                context.line++;
-                context.column = 0;
-            }
-            while (context.column < item.column) {
-                context.result += ' ';
-                context.column++;
-            }
-            if (item instanceof LispContainer) {                
-                item.asText(context);
-            }
-            else {
-                context.result += item.symbol;
-                context.column += item.symbol.length;
-                if (item.isComment() && !item.isLineComment()) {
-                    context.line += item.symbol.split(this.linefeed).length - 1;
-                }
-            }
-        });        
-        return isRoot ? context.result : '';
-    }
-
-}
-
-interface IContainerStringCompilerContext {
-    result: string;
-    line: number;
-    column: number;
-}
-
-
-
-
-
-
+import { Position } from 'vscode';
+import { LispAtom } from './lispAtom';
+
+// This AST Object is a legacy object almost entirely used for formatting purposes
+// Use the Sexpression when you need associated tooling that can preserve whitespace
+// In all other cases, the LispContainer object has much better analytical tooling
 
 enum  LongListFmts{
     kSingleColumn,
@@ -1166,9 +751,9 @@ export class Sexpression extends LispAtom {
     //
     // The formatter has two seperate steps:
     // a. Tokenize: break into the lisp code into individual syntax atoms
-    // b. Assemable the tokens to expected format
+    // b. Assemable the astObjects to expected format
     //
-    // For the second phase it is a recursive algorithm. It layouts the tokens from top to bottom,
+    // For the second phase it is a recursive algorithm. It layouts the astObjects from top to bottom,
     // from left to right. And it follows the bellow rules to handle the individual cases:
     // 1. For the Autolisp builtin APIs if the operator and the first operand are in the same
     //    before formatting, it will use "wide format style". For other cases it uses the "Narrow
