--- extension/src/format/parser.ts
@@ -1,6 +1,0 @@
-// The contents of this file were moved into a folder dedicated to various parsing behaviors.
-// All the existing code integrity was maintained by using this pass through export.
-//    /parsing/lispParser.ts      The original lisp parser mostly used for formatting
-//    /parsing/containers.ts      A different kind of lisp parser with analytical intent
-//    /parsing/comments.ts        Extracts a JSDOC style object from user comments
-export * from '../parsing/lispParser';
