+++ extension/src/project/projectDefinition.ts
@@ -1,4 +1,4 @@
-import { Sexpression } from '../format/sexpression';
+import { Sexpression } from '../astObjects/sexpression';
 import * as nls from 'vscode-nls';
 const localize = nls.config({ messageFormat: nls.MessageFormat.file })();
